refactor: move debug endpoints into separate file
