+++ crates/rpc/src/handlers/block.rs
@@ -1,17 +1,13 @@
-use std::{collections::HashSet, sync::Arc};
-
 use actix_web::{
     HttpRequest, HttpResponse, Responder, get, post,
     web::{Data, Json, Path},
 };
 use alloy_primitives::B256;
-use hashbrown::HashMap;
 use ream_beacon_api_types::{
     error::ApiError,
     id::{ID, ValidatorID},
     responses::{
-        BeaconHeadResponse, BeaconResponse, BeaconVersionedResponse, DataResponse, RootResponse,
-        SSZ_CONTENT_TYPE,
+        BeaconResponse, BeaconVersionedResponse, DataResponse, RootResponse, SSZ_CONTENT_TYPE,
     },
 };
 use ream_consensus::{
@@ -19,9 +15,7 @@ use ream_consensus::{
     electra::{beacon_block::SignedBeaconBlock, beacon_state::BeaconState},
     genesis::Genesis,
 };
-use ream_fork_choice::store::Store;
 use ream_network_spec::networks::network_spec;
-use ream_operation_pool::OperationPool;
 use ream_storage::{
     db::ReamDB,
     tables::{Field, Table},
@@ -298,47 +292,6 @@ pub async fn post_sync_committee_rewards(
     };
 
     Ok(HttpResponse::Ok().json(BeaconResponse::new(reward_data)))
-}
-
-/// Called by `/beacon/heads` to get fork choice leaves.
-#[get("/beacon/heads")]
-pub async fn get_beacon_heads(db: Data<ReamDB>) -> Result<impl Responder, ApiError> {
-    let justified_checkpoint = db.justified_checkpoint_provider().get().map_err(|err| {
-        ApiError::InternalError(format!(
-            "Failed to get justified_checkpoint, error: {err:?}"
-        ))
-    })?;
-
-    let mut blocks = HashMap::new();
-    let store = Store {
-        db: db.get_ref().clone(),
-        operation_pool: Arc::new(OperationPool::default()),
-    };
-
-    store
-        .filter_block_tree(justified_checkpoint.root, &mut blocks)
-        .map_err(|err| {
-            ApiError::InternalError(format!("Failed to filter block tree, error: {err:?}"))
-        })?;
-
-    let mut leaves = vec![];
-    let mut referenced_parents = HashSet::new();
-
-    for block in blocks.values() {
-        referenced_parents.insert(block.parent_root);
-    }
-
-    for (block_root, block) in &blocks {
-        if !referenced_parents.contains(block_root) {
-            leaves.push(BeaconHeadResponse {
-                root: block.block_root(),
-                slot: block.slot,
-                execution_optimistic: false,
-            });
-        }
-    }
-
-    Ok(HttpResponse::Ok().json(DataResponse::new(leaves)))
 }
 
 #[get("/beacon/blind_block/{block_id}")]
