+++ crates/rpc/src/routes/debug.rs
@@ -1,11 +1,7 @@
-use actix_web::web::{ServiceConfig, scope};
+use actix_web::web::ServiceConfig;
 
-use crate::handlers::{block::get_beacon_heads, state::get_beacon_state};
+use crate::handlers::debug::{get_beacon_heads, get_beacon_state};
 
 pub fn register_debug_routes_v2(cfg: &mut ServiceConfig) {
-    cfg.service(
-        scope("/debug")
-            .service(get_beacon_state)
-            .service(get_beacon_heads),
-    );
+    cfg.service(get_beacon_state).service(get_beacon_heads);
 }
